refactor: Move PostTypeIcon into boards dir
