+++ application/millage-client/src/components/boards/PostTypeIcon.tsx
